requirements.txt

apache-airflow[amazon]
apache-airflow-providers-amazon
pandas
marvel
psycopg2-binary





